+++ src/RavenNest.Models/HighScoreItem.cs
@@ -9,7 +9,7 @@ namespace RavenNest.Models
         public string PlayerName { get; set; }
         public int Rank { get; set; }
         public string Skill { get; set; }
-        public decimal Experience { get; set; }
+        public double Experience { get; set; }
         public int Level { get; set; }
     }
 }
