+++ src/RavenNest.SDK/Endpoints/UpdatePlayerExperience.cs
@@ -2,13 +2,13 @@
 {
     public class UpdatePlayerExperience
     {
-        public UpdatePlayerExperience(string userId, decimal[] experience)
+        public UpdatePlayerExperience(string userId, double[] experience)
         {
             UserId = userId;
             Experience = experience;
         }
 
         public string UserId { get; }
-        public decimal[] Experience { get; }
+        public double[] Experience { get; }
     }
 }
